backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
oauthlib==3.3.1
passlib==1.7.4
playwright==1.54.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.37.0
